requirements.txt
beautifulsoup4==4.8.0
bs4==0.0.1
Jinja2==2.10.1
Markdown==3.1.1
MarkupSafe==1.1.1
Pillow==6.1.0
Pygments==2.4.2
soupsieve==1.9.3
